+++ lib/pluck_map/errors.rb
@@ -1,0 +1,4 @@
+module PluckMap
+  class UnsupportedAttributeError < ArgumentError
+  end
+end