+++ test/attribute_builder_test.rb
@@ -3,11 +3,11 @@ require "test_helper"
 class AttributeBuilderTest < Minitest::Test
 
   should "accept either a DSL-style block (no args) or a block that accepts the Builder object" do
-    a = PluckMap::AttributeBuilder.build(model: Author) do
+    a = PluckMap::AttributeBuilder.build(model: Person) do
       last_name
     end
 
-    b = PluckMap::AttributeBuilder.build(model: Author) do |builder|
+    b = PluckMap::AttributeBuilder.build(model: Person) do |builder|
       builder.last_name
     end
 
@@ -15,7 +15,7 @@ class AttributeBuilderTest < Minitest::Test
   end
 
   should "return an array of attributes for each declaration" do
-    attributes = PluckMap::AttributeBuilder.build(model: Author) do
+    attributes = PluckMap::AttributeBuilder.build(model: Person) do
       first_name
       last_name
     end
@@ -24,7 +24,7 @@ class AttributeBuilderTest < Minitest::Test
   end
 
   should "take the declaration as both the column to be selected and the value's name" do
-    attributes = PluckMap::AttributeBuilder.build(model: Author) do
+    attributes = PluckMap::AttributeBuilder.build(model: Person) do
       last_name
     end
 
@@ -33,7 +33,7 @@ class AttributeBuilderTest < Minitest::Test
   end
 
   should "allow overriding the selected attributes with select:" do
-    attributes = PluckMap::AttributeBuilder.build(model: Author) do
+    attributes = PluckMap::AttributeBuilder.build(model: Person) do
       last_name select: :surname
     end
 
@@ -42,7 +42,7 @@ class AttributeBuilderTest < Minitest::Test
   end
 
   should "allow overriding the value's name with as:" do
-    attributes = PluckMap::AttributeBuilder.build(model: Author) do
+    attributes = PluckMap::AttributeBuilder.build(model: Person) do
       last_name as: :lastName
     end
 
@@ -51,7 +51,7 @@ class AttributeBuilderTest < Minitest::Test
   end
 
   should "allow select: to be specified as an array" do
-    attributes = PluckMap::AttributeBuilder.build(model: Author) do
+    attributes = PluckMap::AttributeBuilder.build(model: Person) do
       last_name select: %i{ surname }
     end
 
@@ -61,7 +61,7 @@ class AttributeBuilderTest < Minitest::Test
 
   should "prohibit an empty array for :select" do
     assert_raises ArgumentError do
-      PluckMap::AttributeBuilder.build(model: Author) do
+      PluckMap::AttributeBuilder.build(model: Person) do
         name select: []
       end
     end
@@ -69,23 +69,37 @@ class AttributeBuilderTest < Minitest::Test
 
   should "prohibit multiple selects for :select" do
     assert_raises ArgumentError do
-      PluckMap::AttributeBuilder.build(model: Author) do
+      PluckMap::AttributeBuilder.build(model: Person) do
         name select: %i{ first_name last_name }
       end
     end
   end
 
+  should "prohibit raw SQL for :select" do
+    assert_raises ArgumentError do
+      PluckMap::AttributeBuilder.build(model: Person) do
+        name select: "first_name || ' ' || last_name"
+      end
+    end
+  end
+
+  should "accept Arel::Nodes::SqlLiteral :select" do
+    PluckMap::AttributeBuilder.build(model: Person) do
+      name select: Arel.sql("first_name || ' ' || last_name")
+    end
+  end
+
   should "coerce selects to an empty array of selects when value: is given" do
-    attributes = PluckMap::AttributeBuilder.build(model: Author) do
-      type value: "Author", selects: %i{ a b c }
+    attributes = PluckMap::AttributeBuilder.build(model: Person) do
+      type value: "Person", selects: %i{ a b c }
     end
 
-    assert_equal "Author", attributes[0].value
+    assert_equal "Person", attributes[0].value
     assert_equal [], attributes[0].selects
   end
 
   should "allow value: to be falsey" do
-    attributes = PluckMap::AttributeBuilder.build(model: Author) do
+    attributes = PluckMap::AttributeBuilder.build(model: Person) do
       living value: false
     end
 
@@ -93,7 +107,7 @@ class AttributeBuilderTest < Minitest::Test
   end
 
   should "allow multiple selects when map: is given" do
-    attributes = PluckMap::AttributeBuilder.build(model: Author) do
+    attributes = PluckMap::AttributeBuilder.build(model: Person) do
       name select: %i{ first_name last_name }, map: ->(first, last) { "#{first} #{last}" }
     end