+++ test/support/isbn.rb
@@ -1,0 +1,3 @@
+class Isbn < ActiveRecord::Base
+  belongs_to :book
+end